# relation/openapi.yaml
# Generated with protoc-gen-openapi
# https://github.com/google/gnostic/tree/master/cmd/protoc-gen-openapi

openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
    /douyin/relation/action:
        post:
            tags:
                - FollowAction
            operationId: FollowAction_FollowActionSrv
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/FollowActionRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/FollowActionResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /douyin/relation/follow/list:
        get:
            tags:
                - FollowList
            operationId: FollowList_FollowListSrv
            parameters:
                - name: userId
                  in: query
                  schema:
                    type: integer
                    format: int64
                - name: token
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/FollowListResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /douyin/relation/follower/list:
        get:
            tags:
                - FollowerList
            operationId: FollowerList_FollowerListSrv
            parameters:
                - name: userId
                  in: query
                  schema:
                    type: integer
                    format: int64
                - name: token
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/FollowerListResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /douyin/relation/friend/list:
        get:
            tags:
                - FriendList
            operationId: FriendList_FriendListSrv
            parameters:
                - name: userId
                  in: query
                  schema:
                    type: integer
                    format: int64
                - name: token
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/FriendListResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
components:
    schemas:
        FollowActionRequest:
            type: object
            properties:
                token:
                    type: string
                toUserId:
                    type: integer
                    format: int64
                actionType:
                    type: integer
                    format: int32
        FollowActionResponse:
            type: object
            properties:
                statusCode:
                    type: integer
                    format: int32
                statusMsg:
                    type: string
        FollowListResponse:
            type: object
            properties:
                statusCode:
                    type: integer
                    format: int32
                statusMsg:
                    type: string
                userList:
                    type: array
                    items:
                        $ref: '#/components/schemas/User'
        FollowerListResponse:
            type: object
            properties:
                statusCode:
                    type: integer
                    format: int32
                statusMsg:
                    type: string
                userList:
                    type: array
                    items:
                        $ref: '#/components/schemas/User'
        FriendListResponse:
            type: object
            properties:
                statusCode:
                    type: integer
                    format: int32
                statusMsg:
                    type: string
                userList:
                    type: array
                    items:
                        $ref: '#/components/schemas/User'
        GoogleProtobufAny:
            type: object
            properties:
                '@type':
                    type: string
                    description: The type of the serialized message.
            additionalProperties: true
            description: Contains an arbitrary serialized message along with a @type that describes the type of the serialized message.
        Status:
            type: object
            properties:
                code:
                    type: integer
                    description: The status code, which should be an enum value of [google.rpc.Code][google.rpc.Code].
                    format: int32
                message:
                    type: string
                    description: A developer-facing error message, which should be in English. Any user-facing error message should be localized and sent in the [google.rpc.Status.details][google.rpc.Status.details] field, or localized by the client.
                details:
                    type: array
                    items:
                        $ref: '#/components/schemas/GoogleProtobufAny'
                    description: A list of messages that carry the error details.  There is a common set of message types for APIs to use.
            description: 'The `Status` type defines a logical error model that is suitable for different programming environments, including REST APIs and RPC APIs. It is used by [gRPC](https://github.com/grpc). Each `Status` message contains three pieces of data: error code, error message, and error details. You can find out more about this error model and how to work with it in the [API Design Guide](https://cloud.google.com/apis/design/errors).'
        User:
            type: object
            properties:
                id:
                    type: integer
                    format: int64
                name:
                    type: string
                followCount:
                    type: integer
                    format: int64
                followerCount:
                    type: integer
                    format: int64
                isFollow:
                    type: boolean
                avatar:
                    type: string
                backgroundImage:
                    type: string
                signature:
                    type: string
                totalFavorited:
                    type: integer
                    format: int64
                workCount:
                    type: integer
                    format: int64
                favoriteCount:
                    type: integer
                    format: int64
tags:
    - name: FollowAction
      description: 关注
    - name: FollowList
      description: 关注列表
    - name: FollowerList
      description: 粉丝列表
    - name: FriendList
      description: 朋友列表